pyproject.toml
[project]
name = "customer-service"
version = "0.1.0"
description = "customer service demo using Agent Development Kit"
authors = [{ name = "Christos Aniftos", email = "aniftos@google.com" }]
license = "Apache-2.0"
readme = "README.md"
dependencies = [
    "google-cloud-aiplatform[adk,agent_engines]>=1.93.0",
    "google-adk>=1.0.0",
    "pydantic-settings>=2.8.1",
    "tabulate>=0.9.0",
    "cloudpickle>=3.1.1",
    "pylint>=3.3.6",
    "jsonschema>=4.23.0",
    "pydantic>=2.12.2",
    "langfuse>=3.8.1",
    "openinference-instrumentation-google-adk>=0.1.6",
    "ruff>=0.14.1",
]

requires-python = ">=3.10,<3.13"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-mock>=3.14.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.25.3",
    "flake8-pyproject>=1.2.3",
    "pylint>=3.3.6",
    "pyink>=24.10.1",
    "jsonschema>=4.23.0",
    "google-cloud-aiplatform[evaluation]>=1.93.0",
    "agent-starter-pack>=0.14.1",
]

[project.optional-dependencies]
lint = [
    "ruff>=0.4.6",
    "mypy>=1.15.0",
    "codespell>=2.2.0",
    "types-pyyaml>=6.0.12.20240917",
    "types-requests>=2.32.0.20240914",
]


[tool.pytest.ini_options]
console_output_style = "progress"
addopts = "-vv -s"
#addopts = "-vv -s --pdb"
testpaths = ["tests/", "eval/"]
markers = ["unit"]
log_level = "ERROR"
log_cli = false
log_auto_indent = true
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
log_cli_format = "[%(asctime)s] %(levelname)s (%(funcName)s) \t [%(pathname)s:%(lineno)d] %(message)s"
filterwarnings = ["ignore::UserWarning"]


[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "C",  # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP", # pyupgrade
    "RUF", # ruff specific rules
]
ignore = ["E501", "C901"] # ignore line too long, too complex

[tool.ruff.lint.isort]
known-first-party = ["customer_service"]

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
disallow_subclassing_any = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
follow_imports = "silent"
ignore_missing_imports = true
explicit_package_bases = true
disable_error_code = ["misc", "no-untyped-call", "no-any-return"]

exclude = [".venv"]

[tool.codespell]
ignore-words-list = "rouge"
skip = "./locust_env/*,uv.lock,.venv,./frontend,**/*.ipynb"

[build-system]
requires = ["uv_build>=0.8.14,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = ""

[tool.pyink]
line-length = 80
pyink-indentation = 4
pyink-use-majority-quotes = true

# This configuration file is used by goo.gle/agent-starter-pack to power remote templating.
# It defines the template's properties and settings.
[tool.agent-starter-pack]
example_question = "Hi there, I need help with my garden"

[tool.agent-starter-pack.settings]
agent_directory = "customer_service"
